Split files.js into multiple files
